requirements.txt
streamlit>=1.20.0
openai>=0.27.0
PyPDF2>=3.0.0
python-docx>=0.8.11
rapidfuzz>=2.13.7
python-multipart>=0.0.6
tqdm>=4.65.0
